requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
email-validator
python-dotenv
pydantic-settings
openai
pandas
numpy
scikit-learn
joblib
python-multipart
pytest
pytest-asyncio
pytest-mock
python-jose[cryptography]
cryptography
argon2-cffi
